Updates to use phpunit version 8

All right, in this goes, I hope I'm done with tests for a bit 🥴

+++ .github/setup-database.sh
@@ -25,5 +25,8 @@ if [[ "$DB" == "mysql" ]]; then php .github/SetupMysql.php; fi
 if [[ "$DB" == "mariadb" ]]; then php .github/SetupMysql.php; fi
 if [[ "$DB" == "postgres" ]]; then php .github/SetupPgsql.php; fi
 
+# Show the db install as failed
+if [[ "$?" != 0 ]]; then exit $?; fi
+
 # Remove the install dir
 sudo rm -rf /install
